# Faster k-quants on older GPUs