The oldest GPU I have access to is a GTX-1660 with 6 GB VRAM. The PR significantly improves token prediction on this GPU for `Q2_K - Q4_K` and gets a small boost for `Q5_K` as well (see table). `Q6_K` was not changed. It would be nice to also have some results for the GTX-1000 series (and/or other older GPUs). The PR does not change performance on a modern GPU (with the RTX-4080 that I have available being used for testing). 

| Quantization | t/s master | t/s PR | speed up |
|--:|--:|--:|--:|
| Q2_K | 22.6 | 27.4 | 1.21 |
| Q3_K_S | 17.5 | 24.3 | 1.39 |
| Q4_K_S | 16.3 | 24.4 | 1.50 |
| Q5_K_S | 15.1 | 16.1 | 1.07 |
| Q6_K | 16.5 | 16.5 | 1.00 |
| Q4_0 | 21.4 | - | - |
| Q5_0 | 15.0 | - | - |