@KerfuffleV2 2023-06-18
I tested on a GTX 1060 6GB:

```plaintext
q2_k
master: eval time = 12884.48 ms /    99 runs   (  130.15 ms per token,     7.68 tokens per second)
pull: eval time = 12838.96 ms /    99 runs   (  129.69 ms per token)

q3_k_s
master: eval time = 10787.75 ms /    99 runs   (  108.97 ms per token,     9.18 tokens per second)
pull: eval time = 10813.03 ms /    99 runs   (  109.22 ms per token)

q4_k_s
master: eval time = 18330.58 ms /    99 runs   (  185.16 ms per token,     5.40 tokens per second)
pull: eval time = 18413.52 ms /    99 runs   (  186.00 ms per token)

q5_k_s:
master: eval time = 18408.78 ms /    99 runs   (  185.95 ms per token,     5.38 tokens per second)
pull: eval time = 15824.68 ms /    99 runs   (  159.85 ms per token)
```

TL;DR: q2_k, q3_k, q4_k - no noticeable difference. q5_k actually sped up a decent amount though!