@JohannesGaessler 2023-06-18
My test results:

| GPU      | Test   | Model     | t/s master | t/s PR | t/s PR LLAMA_CUDA_KQUANTS_ITER=1 | Speedup |
|----------|--------|-----------|------------|--------|----------------------------------|---------|
| GTX 1070 | TG 128 | 7b Q2_K   |       8.80 |   8.75 |                            11.11 |    1.26 |
| GTX 1070 | TG 128 | 7b Q3_K_S |       9.15 |   9.05 |                             9.78 |    1.07 |
| GTX 1070 | TG 128 | 7b Q4_K_S |       6.61 |   6.60 |                             8.79 |    1.33 |
| GTX 1070 | TG 128 | 7b Q5_K_S |       6.15 |   6.56 |                             6.56 |    1.07 |
| GTX 1070 | TG 128 | 7b Q4_0   |      10.23 |      - |                                - |       - |
| GTX 1070 | TG 128 | 7b Q4_0   |       7.07 |      - |                                - |       - |