@KerfuffleV2 2023-06-18
Wow, `LLAMA_CUDA_KQUANTS_ITER` makes a huge difference on my 1060, at least for q4_k.

```plaintext
Q5_K_S: eval time = 15554.61 ms /    99 runs   (  157.12 ms per token)
Q4_K_S: eval time = 10581.56 ms /    99 runs   (  106.88 ms per token)
Q3_K_S: eval time =  9499.68 ms /    99 runs   (   95.96 ms per token)
Q2_K: eval time =  8531.30 ms /    99 runs   (   86.17 ms per token)
```

Nearly doubled Q4_K speed, and a nice bump for Q2_K for anyone crazy enough to actually try to use it.

I also tested with `LLAMA_CUDA_KQUANTS_ITER=1` on master:

```plaintext
eval time = 16926.22 ms /    99 runs   (  170.97 ms per token,     5.85 tokens per second)
```

It's a little faster, but nothing dramatic, so it's the combination of this PR and that tweak which makes the difference. Sadly, I get the best token generation speed of all by not offloading anything (~84ms/token) but people with old GPUs and worse CPUs will definitely see a lot of benefit from the change.