@johnson442 2023-06-19
LLAMA_CUDA_KQUANTS_ITER=1 on master and PR with a 1070ti:
|	|Master|	|PR	|	|
|-------|------|--------|-------|-------|	
|7B	|ms/t	|t/s	|ms/t	|t/s	|
|Q2_K	|73.66	|13.58	|57.3	|17.45	|
|Q3_K_S	|70.62	|14.16	|60.7	|16.47	|
|Q4_K_S	|117.94	|8.48	|73.34	|13.64	|
|Q5_K_S	|122.87	|8.14	|117.29	|8.53	|
|Q6_K	|70.32	|14.22	|70.31	|14.22	|
|	|	|	|	|	|
|13B Q3_K_S (-ngl 42 -t 5)|150.96	|6.62	|118.31	|8.45|
|30B Q3_K_S (-ngl 28 -t 5)|357.3	|2.80	|334.31	|2.99|

I'm off to download some Q4_K models!