@JohannesGaessler 2023-06-19
I think the table that I just added in README.md regarding compilation options will need to be adapted but otherwise I think this is good to merge.